Implement the authorization

+++ app/Http/Controllers/IdeaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Idea;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class IdeaController extends Controller
 {
@@ -12,7 +13,8 @@ class IdeaController extends Controller
         $validated = $request->validate([
             'content' => ['required', 'min:3', 'max:240']
         ]);
-        Idea::created($validated);
+        $validated['user_id'] = Auth::id();
+        Idea::create($validated);
         return redirect()->route('home')->with('success', 'Idea was created successfully');
 
 
@@ -20,6 +22,9 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if (Auth::id() !== $idea->user_id) {
+            abort(404);
+        }
         $idea->delete();
 
         return redirect()->route('home')->with('success', 'The idea successfully deleted');
@@ -32,12 +37,18 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (Auth::id() !== $idea->user_id) {
+            abort(404);
+        }
         $editing = true;
         return view('ideas.show', compact('idea', 'editing'));
     }
 
     public function update(Request $request, Idea $idea)
     {
+        if (Auth::id() !== $idea->user_id) {
+            abort(404);
+        }
         $validated = $request->validate([
             'content' => ['required', 'min:3', 'max:240']
         ]);
